refactor: suppress warnings

+++ src/iceberg/util/truncate_util.cc
@@ -66,7 +66,7 @@ Literal TruncateLiteralImpl<TypeId::kBinary>(const Literal& literal, int32_t wid
   // In contrast to strings, binary values do not have an assumed encoding and are
   // truncated to `width` bytes.
   const auto& data = std::get<std::vector<uint8_t>>(literal.value());
-  if (data.size() <= width) {
+  if (data.size() <= static_cast<size_t>(width)) {
     return literal;
   }
   return Literal::Binary(std::vector<uint8_t>(data.begin(), data.begin() + width));
